+++ src/Function/ArgMapFunction.php
@@ -5,7 +5,7 @@
 namespace MediaWiki\Extension\ParserPower\Function;
 
 use MediaWiki\Extension\ParserPower\Operation\TemplateOperation;
-use MediaWiki\Extension\ParserPower\ParameterParser;
+use MediaWiki\Extension\ParserPower\Parameters;
 use MediaWiki\Extension\ParserPower\ParserPower;
 use MediaWiki\Parser\Parser;
 use MediaWiki\Parser\PPFrame;
@@ -46,7 +46,7 @@ final class ArgMapFunction extends ParserFunctionBase {
 	/**
 	 * @inheritDoc
 	 */
-	public function execute( Parser $parser, PPFrame $frame, ParameterParser $params ): string {
+	public function execute( Parser $parser, PPFrame $frame, Parameters $params ): string {
 		// set parameters
 		$formatter = $params->get( 'formatter' );
 		if ( $formatter === '' ) {
